fix clustering
